Empathy map
Empathy Map:
The overall user analysis is shown here descriptively with all the possible constraints that directly impact the user. Based on these constraints , the user interface design will be created and enhance the potential productivity of the project. This empathy map depicts user point of view in all aspects of conditionals available .